+++ app/services/bill-runs/supplementary/process-billing-period.service.js
@@ -10,8 +10,8 @@ const BillRunModel = require('../../../models/bill-run.model.js')
 const BillModel = require('../../../models/bill.model.js')
 const BillLicenceModel = require('../../../models/bill-licence.model.js')
 const DetermineChargePeriodService = require('../determine-charge-period.service.js')
-const DetermineMinimumChargeService = require('./determine-minimum-charge.service.js')
-const GenerateTransactionsService = require('./generate-transactions.service.js')
+const DetermineMinimumChargeService = require('../determine-minimum-charge.service.js')
+const GenerateTransactionsService = require('../generate-transactions.service.js')
 const PreGenerateBillingDataService = require('./pre-generate-billing-data.service.js')
 const ProcessTransactionsService = require('./process-transactions.service.js')
 const SendTransactionsService = require('./send-transactions.service.js')
