+++ test/services/bill-runs/determine-minimum-charge.service.test.js
@@ -8,12 +8,12 @@ const { describe, it, beforeEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Test helpers
-const ChangeReasonHelper = require('../../../support/helpers/change-reason.helper.js')
-const ChargeVersionHelper = require('../../../support/helpers/charge-version.helper.js')
-const DatabaseHelper = require('../../../support/helpers/database.helper.js')
+const ChangeReasonHelper = require('../../support/helpers/change-reason.helper.js')
+const ChargeVersionHelper = require('../../support/helpers/charge-version.helper.js')
+const DatabaseHelper = require('../../support/helpers/database.helper.js')
 
 // Thing under test
-const DetermineMinimumChargeService = require('../../../../app/services/bill-runs/supplementary/determine-minimum-charge.service.js')
+const DetermineMinimumChargeService = require('../../../app/services/bill-runs/determine-minimum-charge.service.js')
 
 describe('Determine Minimum Charge service', () => {
   const chargePeriod = {
